various updates; see details
• reworked kill tracking in that even if no regex match kill starts the
fight timer reset
• removed dispatcher on log publishing as it locks gui
• added json validation check before adding any entries into lists for
saving
• updated mapindex to UInt32
• added setting and local options to allow displaying of help labels;
turning off will minimize size of gui
• updated defeats regex for all language but German
• updated common library with new settings for nlog and labels

+++ FFXIVAPP.Client/Memory/NPCEntry.cs
@@ -11,7 +11,7 @@ namespace FFXIVAPP.Client.Memory
     {
         #region Property Backings
 
-        public ushort MapIndex { get; set; }
+        public uint MapIndex { get; set; }
         public string Name { get; set; }
         public uint ID { get; set; }
         public uint NPCID { get; set; }
